+++ packages/deprecation-crawler/src/lib/output-formatters/index.ts
@@ -1,5 +1,5 @@
 export {addCommentToRepository} from './add-comments-to-repository';
-export {generateRawJson} from './generate-raw-json';
+export {generateRawJson} from './json/raw.json.formatter';
 import {generateTagBasedFormat} from './markdown/tag-based.md-formatter';
 import {generateGroupBasedFormat} from './markdown/group-based.md.formatter';
